move wizardState into types/ folder
